local Data Flow Server restructuring
- spring-cloud-dataflow-deployer-local implements the deployer SPI for local execution
- spring-cloud-starter-dataflow-server-local brings the local deployer and the dataflow-server-core dependencies together
- spring-cloud-dataflow-server-local depends on the starter above and provides the main class: LocalDataFlowServer

removed "module" from spring-cloud-stream-module-deployer-spi
