--- spring-cloud-dataflow-admin-local/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  org.springframework.cloud.dataflow.admin.spi.local.LocalDeployerAutoConfiguration
