+++ spring-cloud-dataflow-dependencies/pom.xml
@@ -46,7 +46,12 @@
 			</dependency>
 			<dependency>
 				<groupId>org.springframework.cloud</groupId>
-				<artifactId>spring-cloud-dataflow-module-deployer-spi</artifactId>
+				<artifactId>spring-cloud-dataflow-deployer-spi</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>org.springframework.cloud</groupId>
+				<artifactId>spring-cloud-dataflow-deployer-local</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 			<dependency>
@@ -61,7 +66,7 @@
 			</dependency>
 			<dependency>
 				<groupId>org.springframework.cloud</groupId>
-				<artifactId>spring-cloud-dataflow-admin-local</artifactId>
+				<artifactId>spring-cloud-dataflow-server-local</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 			<dependency>
