+++ spring-cloud-dataflow-deployer-local/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+  org.springframework.cloud.dataflow.deployer.local.LocalDeployerAutoConfiguration
