+++ spring-cloud-dataflow-server-core/pom.xml
@@ -31,7 +31,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-dataflow-module-deployer-spi</artifactId>
+			<artifactId>spring-cloud-dataflow-deployer-spi</artifactId>
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
