+++ spring-cloud-dataflow-server-local/README.adoc
@@ -1,0 +1,131 @@
+== Spring Cloud Data Flow Server - Local
+
+The SPI implementation for deploying https://github.com/spring-cloud/spring-cloud-stream[Spring Cloud Stream] modules locally, either within the same JVM as the Data Flow Server or by spawning a new JVM process on the local machine. Spawning a new JVM process is the default. To run modules within the existing JVM process set the environment variable `DEPLOYER_LOCAL_OUT_OF_PROCESS=false` or the corresponding property: `deployer.local.out-of-process`.
+
+== Getting Started
+
+This shows how to start the Data Flow Server and Shell to create the `time | log` stream.
+
+=== Start the Data Flow Server and Shell
+
+. Start Redis locally via `redis-server`
+. Build from the spring-cloud-dataflow root directory:
++
+----
+./mvnw clean install
+----
++
+. Start the Data Flow Server application:
++
+----
+export DEPLOYER_LOCAL_OUT_OF_PROCESS=false
+
+java -jar spring-cloud-dataflow-server-local/target/spring-cloud-dataflow-server-local-1.0.0.BUILD-SNAPSHOT.jar
+----
++
+. Start the shell
++
+----
+$ java -jar spring-cloud-dataflow-shell/target/spring-cloud-dataflow-shell-1.0.0.BUILD-SNAPSHOT.jar
+----
+
+=== Creating the `time | log` stream:
+
+. Create the 'ticktock' stream:
++
+----
+dataflow:>stream create --name ticktock --definition "time | log"
+Created new stream 'ticktock'
+----
++
+This is equivalent to the following HTTP POST request
++
+----
+$ curl -X POST -d "name=ticktock&definition=time | log" http://localhost:9393/streams/definitions?deploy=false
+----
++
+. List all streams available in the repository:
++
+----
+dataflow:>stream list
+╔═══════════╤═════════════════╤══════════╗
+║Stream Name│Stream Definition│  Status  ║
+╠═══════════╪═════════════════╪══════════╣
+║ticktock   │time | log       │undeployed║
+╚═══════════╧═════════════════╧══════════╝
+----
++
+This is equivalent to the following HTTP get request
++
+----
+$ curl http://localhost:9393/streams/definitions
+----
++
+. Deploy the 'ticktock' stream:
++
+----
+dataflow:>stream deploy --name ticktock
+Deployed stream 'ticktock'
+----
++
+This is equivalent to the following HTTP get request
++
+----
+$ curl -X POST http://localhost:9393/streams/deployments/ticktock
+----
+
+If successful you should see output similar to the following in the `Data Flow Server` console:
+
+----
+2015-12-08 16:36:23.917  INFO 22951 --- [hannel-adapter1] log.sink  : 2015-12-08 16:36:23
+2015-12-08 16:36:24.918  INFO 22951 --- [hannel-adapter1] log.sink  : 2015-12-08 16:36:24
+2015-12-08 16:36:25.919  INFO 22951 --- [hannel-adapter1] log.sink  : 2015-12-08 16:36:25
+----
+
+## Configuration
+
+### Default
+To configure the Data Flow Server you can follow the configuration setup guidelines specified in the boot documentation found http://docs.spring.io/spring-boot/docs/current/reference/html/boot-features-external-config.html[here]
+
+Note: The `admin.yml` containing the defaults can be found https://github.com/spring-cloud/spring-cloud-dataflow/blob/master/spring-cloud-dataflow-deployer-local/src/main/resources/admin.yml[here]
+
+### Spring Cloud Configuration
+The Spring Cloud Data Flow Server offers the user the ability to configure properties via
+http://cloud.spring.io/spring-cloud-config/spring-cloud-config.html[spring-cloud-config].
+All configurations retrieved from the cloud config will take precedence over Boot's
+defaults enumerated above. The Spring Cloud Data Flow Server will look for the server at
+`localhost:8888`, however this can be overwritten by setting the `spring.cloud.config.uri`
+property to the desired url.
+
+#### Cloud-Config-Server configuration
+
+To specify a repository in the cloud config server configuration.yml for the Data Flow Server,
+setup a repo profile with the pattern `spring-cloud-dataflow-server`. For example:
+
+[source,yml]
+----
+spring:
+  cloud:
+     config:
+       server:
+         git:
+           uri: https://github.com/myrepo/configurations
+           repos:
+            spring-cloud-dataflow-server:
+              pattern: spring-cloud-dataflow-server
+              uri: https://github.com/myrepo/configurations
+              searchPaths: dataFlowServer
+----
+
+==== Fail Fast
+In some cases, it may be desirable to fail startup of a service if it cannot connect to
+the Config Server. If this is the desired behavior, set the bootstrap configuration
+property `spring.cloud.config.failFast=true` and the client will halt with an Exception.
+
+==== Note
+If the Data Flow Server cannot connect to the cloud config server, the
+following warning message will be logged: 
+----
+`WARN 42924 --- [           main] c.c.c.ConfigServicePropertySourceLocator : Could not locate PropertySource: I/O error on GET request for "http://localhost:8888/spring-cloud-dataflow-server/default":Connection refused; nested exception is java.net.ConnectException: Connection refused`
+----
+To disable the cloud config server set the `spring.cloud.config.enabled` property to `false`.
