+++ spring-cloud-dataflow-server-local/src/main/java/org/springframework/cloud/dataflow/server/local/LocalDataFlowServer.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2016 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.cloud.dataflow.server.local;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.cloud.dataflow.admin.EnableDataFlowServer;
+
+/**
+ * Bootstrap class for the local Spring Cloud Data Flow Server.
+ *
+ * @author Mark Fisher
+ */
+@EnableDataFlowServer
+public class LocalDataFlowServer {
+
+	public static void main(String[] args) {
+		SpringApplication.run(LocalDataFlowServer.class, args);
+	}
+
+}
