+++ spring-cloud-starter-dataflow-server-local/pom.xml
@@ -17,7 +17,7 @@
 	<dependencies>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-dataflow-admin-local</artifactId>
+			<artifactId>spring-cloud-dataflow-deployer-local</artifactId>
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
